Implement smart author detection
Add a mechanism of more complicated string comparison while parsing
occurs. When the Title field contains symbol '#' converter begins to
analyze text after that symbol as the Author label and tries to find
that label in the name of parsing source title.

Also I have added the tip message in the Add and Edit windows about that
new symbol